pyproject.toml
[project]
name = "tradefarm"
version = "0.1.0"
description = "100-agent AI trading platform for US stocks/ETFs"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "alpaca-py>=0.33",
    "anthropic>=0.40",
    "pandas>=2.2",
    "pyarrow>=17.0",
    "numpy>=2.0",
    "pandas-market-calendars>=4.4",
    "python-dotenv>=1.0",
    "structlog>=24.4",
    "websockets>=13.1",
]

[project.optional-dependencies]
ml = [
    "torch>=2.5",
    "scikit-learn>=1.5",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
    "mypy>=1.11",
    "pillow>=12.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tradefarm"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
